Refactor defaults into options
