+++ src/index.js
@@ -2,10 +2,14 @@
 
 module.exports = {
   BasePaginator: require('./structures/BasePaginator'),
-  ButtonPaginator: require('./structures/ButtonPaginator'),
-  ActionRowPaginator: require('./structures/ActionRowPaginator'),
-  Defaults: require('./util/Defaults'),
-  PaginatorEvents: require('./util/PaginatorEvents'),
+  BaseOptions: require('./util/BaseOptions'),
   ReactionPaginator: require('./structures/ReactionPaginator'),
+  ReactionPaginatorOptions: require('./util/ReactionPaginatorOptions'),
+  ActionRowPaginator: require('./structures/ActionRowPaginator'),
+  ActionRowPaginatorOptions: require('./util/ActionRowPaginatorOptions'),
+  ButtonPaginator: require('./structures/ButtonPaginator'),
+  ButtonPaginatorOptions: require('./util/ButtonPaginatorOptions'),
   SelectPaginator: require('./structures/SelectPaginator'),
+  SelectPaginatorOptions: require('./util/SelectPaginatorOptions'),
+  PaginatorEvents: require('./util/PaginatorEvents'),
 };
