+++ src/structures/ReactionPaginator.js
@@ -2,11 +2,11 @@
 
 const { Util } = require('discord.js');
 const BasePaginator = require('./BasePaginator');
-const { ReactionPaginatorDefaults } = require('../util/Defaults');
+const ReactionPaginatorOptions = require('../util/ReactionPaginatorOptions');
 
 class ReactionPaginator extends BasePaginator {
   constructor(interaction, options) {
-    super(interaction, Util.mergeDefault(ReactionPaginatorDefaults, options));
+    super(interaction, Util.mergeDefault(ReactionPaginatorOptions.createDefault(), options));
 
     if (typeof options.emojiList === 'undefined' || options.emojiList.length === 0) {
       throw new Error('emojiList is undefined or empty, must be a list of EmojiResolvables');
