+++ src/structures/SelectPaginator.js
@@ -2,11 +2,11 @@
 
 const { Util } = require('discord.js');
 const ActionRowPaginator = require('./ActionRowPaginator');
-const { SelectPaginatorDefaults } = require('../util/Defaults');
+const SelectPaginatorOptions = require('../util/SelectPaginatorOptions');
 
 class SelectPaginator extends ActionRowPaginator {
   constructor(interaction, options) {
-    super(interaction, Util.mergeDefault(SelectPaginatorDefaults, options));
+    super(interaction, Util.mergeDefault(SelectPaginatorOptions.createDefault(), options));
     if (this.options.selectOptions) {
       const actionRows = [[]];
       this.options.selectOptions.forEach(selectOption => {
